Save both the initial and the refined wavelength calibration.
There is no need to use a negative value of the polynomial degree to skip the refined calibration. Now the code saves both calibrations. The initial one is stored in initial_master_wlcalib.json. The refined version is stored in master_wlcalib.json. Note that when using line lists in the old format, only the master_wlcalib.json will be generated, using in this case the preliminary version of the wavelength calibration.